Stitching POC is working
Now for the real deal - blending a mesh onto an infinitely huge canvas.

The current stitcher "pulls" pixels from the source texture (ie implicit
rendering), but I want to switch to an explicit model, where the new
frame is rendered as a mesh, on top of the infinitely large canvas,
and we need to switch to GPU rendering for that now.