Removed horz drift lock, fixed first frame horz bias
This is much better.

1. Discovered that we're aligning our horizontal drift to the bottom
center grid vertex, and I had incorrectly assumed that this vertex
would lie at the center of the image. It only lies at the center of
the image if there is an odd number of grid columns. We now ensure
that the number of columns is odd, and this fixes that strange
initial, large horizontal drift.

2. Fixing this has allowed me to remove all of the hacks that were
in place to try to keep the alignment proceeding perfectly
horizontally. We now drift freely, and the results look better.