Wrap up new stitcher into a class called VideoStitcher
I have built a new speed computer, based on this new VideoStitcher.
That speed computer is Speed3, and it's much simpler, now that
the stitching has been abstracted into VideoStitcher. I am busy
producing Stitcher3, which will be a similar attempt, to change
the giant image stitcher to use the new abstraction too.